Addressing PR comments.
